pyproject.toml
[project]
name = "shareit-act"
version = "1.0.0"
description = "To comply with H.R.9566 - SHARE IT Act and share open source code with other federal agencies"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pygithub>=2.6.1",
    "python-dotenv>=1.1.0",
]
